STT re-organized
