+++ stt/all_benchmarks.csv
@@ -1,28 +1,28 @@
 provider,model,voice,benchmark,metric_type,metric_value,metric_units,audio_filename,timestamp,status
-deepgram,nova-2,N/A,STT,TTFT,0.22933030128479004,s,common_voice_sample_1.wav,2025-08-05 10:35:10,success
-deepgram,nova-2,N/A,STT,WER,0.0,%,common_voice_sample_1.wav,2025-08-05 10:35:10,success
-deepgram,nova-2,N/A,STT,RTF,0.9400764812997282,,common_voice_sample_1.wav,2025-08-05 10:35:10,success
-deepgram,nova-3,N/A,STT,TTFT,0.22844696044921875,s,common_voice_sample_1.wav,2025-08-05 10:35:10,success
-deepgram,nova-3,N/A,STT,WER,0.0,%,common_voice_sample_1.wav,2025-08-05 10:35:10,success
-deepgram,nova-3,N/A,STT,RTF,0.935063905834722,,common_voice_sample_1.wav,2025-08-05 10:35:10,success
-assemblyai,universal-streaming,N/A,STT,TTFT,1.5591316223144531,s,common_voice_sample_1.wav,2025-08-05 10:35:10,success
-assemblyai,universal-streaming,N/A,STT,WER,0.0,%,common_voice_sample_1.wav,2025-08-05 10:35:10,success
-assemblyai,universal-streaming,N/A,STT,RTF,0.7922662830150217,,common_voice_sample_1.wav,2025-08-05 10:35:10,success
-speechmatics,default,N/A,STT,TTFT,0.9269065856933594,s,common_voice_sample_1.wav,2025-08-05 10:35:10,success
-speechmatics,default,N/A,STT,WER,0.0,%,common_voice_sample_1.wav,2025-08-05 10:35:10,success
-speechmatics,default,N/A,STT,RTF,0.8729742615343676,,common_voice_sample_1.wav,2025-08-05 10:35:10,success
-speechmatics,enhanced,N/A,STT,TTFT,1.082317590713501,s,common_voice_sample_1.wav,2025-08-05 10:35:10,success
-speechmatics,enhanced,N/A,STT,WER,0.0,%,common_voice_sample_1.wav,2025-08-05 10:35:10,success
-speechmatics,enhanced,N/A,STT,RTF,0.8407891493325356,,common_voice_sample_1.wav,2025-08-05 10:35:10,success
-google,short,N/A,STT,TTFT,1.9407577514648438,s,common_voice_sample_1.wav,2025-08-05 10:35:10,success
-google,short,N/A,STT,WER,0.0,%,common_voice_sample_1.wav,2025-08-05 10:35:10,success
-google,short,N/A,STT,RTF,0.863591313956378,,common_voice_sample_1.wav,2025-08-05 10:35:10,success
-google,long,N/A,STT,TTFT,1.7220790386199951,s,common_voice_sample_1.wav,2025-08-05 10:35:10,success
-google,long,N/A,STT,WER,0.0,%,common_voice_sample_1.wav,2025-08-05 10:35:10,success
-google,long,N/A,STT,RTF,0.9028753826869219,,common_voice_sample_1.wav,2025-08-05 10:35:10,success
-google,telephony,N/A,STT,TTFT,1.776944637298584,s,common_voice_sample_1.wav,2025-08-05 10:35:10,success
-google,telephony,N/A,STT,WER,0.0,%,common_voice_sample_1.wav,2025-08-05 10:35:10,success
-google,telephony,N/A,STT,RTF,0.8738586579525217,,common_voice_sample_1.wav,2025-08-05 10:35:10,success
-google,chirp_2,N/A,STT,TTFT,1.6229174137115479,s,common_voice_sample_1.wav,2025-08-05 10:35:10,success
-google,chirp_2,N/A,STT,WER,0.0,%,common_voice_sample_1.wav,2025-08-05 10:35:10,success
-google,chirp_2,N/A,STT,RTF,0.8175727822253083,,common_voice_sample_1.wav,2025-08-05 10:35:10,success
+deepgram,nova-2,N/A,STT,TTFT,0.7707431316375732,s,common_voice_sample_1.wav,2025-08-05 20:39:42,success
+deepgram,nova-2,N/A,STT,WER,0.0,%,common_voice_sample_1.wav,2025-08-05 20:39:42,success
+deepgram,nova-2,N/A,STT,RTF,0.9758415309413486,,common_voice_sample_1.wav,2025-08-05 20:39:42,success
+deepgram,nova-3,N/A,STT,TTFT,0.7757482528686523,s,common_voice_sample_1.wav,2025-08-05 20:39:42,success
+deepgram,nova-3,N/A,STT,WER,0.0,%,common_voice_sample_1.wav,2025-08-05 20:39:42,success
+deepgram,nova-3,N/A,STT,RTF,0.9748628176188581,,common_voice_sample_1.wav,2025-08-05 20:39:42,success
+assemblyai,universal-streaming,N/A,STT,TTFT,2.414102077484131,s,common_voice_sample_1.wav,2025-08-05 20:39:42,success
+assemblyai,universal-streaming,N/A,STT,WER,0.0,%,common_voice_sample_1.wav,2025-08-05 20:39:42,success
+assemblyai,universal-streaming,N/A,STT,RTF,0.7309120130186432,,common_voice_sample_1.wav,2025-08-05 20:39:42,success
+speechmatics,default,N/A,STT,TTFT,1.779120683670044,s,common_voice_sample_1.wav,2025-08-05 20:39:42,success
+speechmatics,default,N/A,STT,WER,0.0,%,common_voice_sample_1.wav,2025-08-05 20:39:42,success
+speechmatics,default,N/A,STT,RTF,0.8584044551337456,,common_voice_sample_1.wav,2025-08-05 20:39:42,success
+speechmatics,enhanced,N/A,STT,TTFT,1.9409611225128174,s,common_voice_sample_1.wav,2025-08-05 20:39:42,success
+speechmatics,enhanced,N/A,STT,WER,0.0,%,common_voice_sample_1.wav,2025-08-05 20:39:42,success
+speechmatics,enhanced,N/A,STT,RTF,0.8140538718302963,,common_voice_sample_1.wav,2025-08-05 20:39:42,success
+google,short,N/A,STT,TTFT,2.3854875564575195,s,common_voice_sample_1.wav,2025-08-05 20:39:42,success
+google,short,N/A,STT,WER,8.333333333333332,%,common_voice_sample_1.wav,2025-08-05 20:39:42,success
+google,short,N/A,STT,RTF,0.836081459359784,,common_voice_sample_1.wav,2025-08-05 20:39:42,success
+google,long,N/A,STT,TTFT,2.081573009490967,s,common_voice_sample_1.wav,2025-08-05 20:39:42,success
+google,long,N/A,STT,WER,0.0,%,common_voice_sample_1.wav,2025-08-05 20:39:42,success
+google,long,N/A,STT,RTF,0.8572793829230528,,common_voice_sample_1.wav,2025-08-05 20:39:42,success
+google,telephony,N/A,STT,TTFT,2.380390167236328,s,common_voice_sample_1.wav,2025-08-05 20:39:42,success
+google,telephony,N/A,STT,WER,0.0,%,common_voice_sample_1.wav,2025-08-05 20:39:42,success
+google,telephony,N/A,STT,RTF,0.6456479478788726,,common_voice_sample_1.wav,2025-08-05 20:39:42,success
+google,chirp_2,N/A,STT,TTFT,2.078447103500366,s,common_voice_sample_1.wav,2025-08-05 20:39:42,success
+google,chirp_2,N/A,STT,WER,0.0,%,common_voice_sample_1.wav,2025-08-05 20:39:42,success
+google,chirp_2,N/A,STT,RTF,0.7655091257327595,,common_voice_sample_1.wav,2025-08-05 20:39:42,success
